fix: address ZeroOS security review findings
Soundness fixes:
- P-256 ECDSA: reject trivial zero GLV decomposition (a=0, b=0) that
  collapses the Shamir MSM and allows signature forgery for any message
  and public key. secp256k1 is not affected.
- rd=x0 allowlist: exclude SCW/SCD/LRW/LRD/INLINE from the generic
  rd=x0 trace rewriting path so their custom trace() methods run,
  preserving VirtualAdvice patching and reservation state management.

Completeness fix:
- EBREAK: expand into a JAL-to-self virtual sequence so the Jump flag
  disables the NextUnexpPCUpdateOtherwise constraint at the NoOp
  padding boundary.

Emulator fix:
- MULHSU exec(): sign-extend rs1 via `as i128 as u128` instead of
  zero-extending via `as u128`. Only affects non-tracing emulator mode;
  proof generation uses the correct inline_sequence path.

Documentation:
- Document the M-mode-only privilege and trap model in the book and in
  ecall.rs/mret.rs code comments. ECALL's constant mstatus write and
  MRET's omitted mstatus restoration are correct by design for the
  single-privilege, no-interrupt execution model.

+++ tracer/src/instruction/mulhsu.rs
@@ -31,7 +31,7 @@ impl MULHSU {
                         >> 32,
                 ),
                 Xlen::Bit64 => {
-                    ((cpu.x[self.operands.rs1 as usize] as u128)
+                    ((cpu.x[self.operands.rs1 as usize] as i128 as u128)
                         .wrapping_mul(cpu.x[self.operands.rs2 as usize] as u64 as u128)
                         >> 64) as i64
                 }
